requirements.txt
python-telegram-bot==21.0
groq
httpx[http2]
certifi